fix count type

+++ HomeCook/Models/Position.swift
@@ -1,9 +1,9 @@
 
 struct Position {
     let dish: Dish
-    var count: Double
+    var count: Int
     let cook: Cook
-    var cost: Double {
+    var cost: Int {
         dish.price * count
     }
 }
